rework based on comments
